slate-satellite-python/requirements.txt
boto3==1.16.14
Flask
pyyaml
zstandard==0.14.0
requests